requirements.txt
chromadb==0.4.22
google-generativeai==0.3.2
python-dotenv==1.0.0
uuid
logging
